# prototipo/static/swagger/api.json
{
    "swagger": "2.0",
    "info": {
        "title": "Uchuva API",
        "description": "DAG manager",
        "version": "1.0.0"
    },
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "consumes": [
    "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
        "/listar": {
            "post": {
                "summary": "List files and folders",
                "description": "The listar endpoint returns a list of files and folders from an user\n",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "path of the folder",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/pathcwd"
                        }
                    }
                ],
                "tags": [
                    "VFS"
                ],
                "responses": {
                    "200": {
                        "description": "Files of the user",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "result": {
                                    "type": "array",
                                    "description": "array of files",
                                    "items": {
                                      "$ref": "#/definitions/Fileview"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/crearArchivo": {
            "post": {
                "summary": "Create file",
                "description": "The crearArchivos endpoint save a file in the user fs.\nThe response includes a result of the operation\n",
                "consumes": [
                  "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cwd",
                        "in": "formData",
                        "description": "Identifier of the dag.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "File",
                        "required": true,
                        "type": "file"
                    }
                ],
                "tags": [
                    "VFS"
                ],
                "responses": {
                    "200": {
                        "description": "Result of the operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "string",
                                    "description": "Error number. None is 0"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/descargarArchivo": {
            "get": {
                "summary": "Download a file",
                "description": "Download a sigle file",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "path",
                        "in": "query",
                        "description": "Indentifier of the file",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "VFS"
                ],
                "responses": {
                    "200": {
                        "description": "Name and identifier of the dag",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/eliminarArchivo": {
            "post": {
                "summary": "Delete files",
                "description": "The eliminarArchivo endpoint deletes a set of files from vfs.\nThe response includes the result code of the operation\n",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                      "name": "item",
                      "in": "body",
                      "description": "Id of the file",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                    }
                ],
                "tags": [
                    "VFS"
                ],
                "responses": {
                    "200": {
                        "description": "Result of the operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "number",
                                    "description": "Id of the file"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/contenidoArchivo": {
            "post": {
                "summary": "Get the content of the file",
                "description": "The contenidoArchivo endpoint getst the content of a file from vfs.\nThe response includes the content of the file\n",
                "consumes": [
                  "application/json"
                ],
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                      "name": "item",
                      "in": "body",
                      "description": "Id of the file",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                    }
                ],
                "tags": [
                    "VFS"
                ],
                "responses": {
                    "200": {
                        "description": "Content of the file",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "result": {
                                    "type": "string",
                                    "description": "Content of the file"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/crearCarpeta": {
            "post": {
                "summary": "Create a folder in the VFS",
                "description": "The crearCarpeta endpoint create a folder in the vfs.\nThe response includes a result code of the operation path\n",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "newPath",
                        "in": "body",
                        "description": "item to rename",
                        "schema": {
                          "$ref": "#/definitions/Folder"
                        }
                    }
                ],
                "tags": [
                    "VFS"
                ],
                "responses": {
                    "200": {
                        "description": "The result of the operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "number",
                                    "description": "result of create folder"
                                },
                                "id": {
                                    "type": "string",
                                    "description": "result of create folder"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pathcwd": {
            "type": "object",
            "properties": {
                "cwd": {
                    "type": "string",
                    "description": "Position in pagination."
                }
            }
        },
        "Folder": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "cwd": {
                    "type": "string"
                }
            }
        },
        "Fileview": {
          "type": "object",
          "properties": {
            "name":{
                "type": "string"
            },
            "rights":{
                "type": "string"
            },
            "date":{
                "type": "string"
            },
            "type":{
                "type": "string"
            },
            "id":{
                "type": "string"
            }
          }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}